# MediClaim.API/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MediClaim API",
    "description": "Enterprise Claims API",
    "contact": {
      "name": "MediClaim Team",
      "email": "support@mediclaim.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:44312",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/policies/types": {
      "post": {
        "tags": [
          "Policies"
        ],
        "operationId": "createPolicyType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyTypeCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyTypeCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyTypeCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "get": {
        "tags": [
          "Policies"
        ],
        "operationId": "getPolicyTypes",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/policies/enroll": {
      "post": {
        "tags": [
          "Policies"
        ],
        "summary": "Enroll patient into policy.",
        "description": "Creates a new policy enrollment for the specified patient.",
        "operationId": "enroll",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollPolicyCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollPolicyCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollPolicyCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Policy enrolled successfully."
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/policies/{policyId}/upgrade": {
      "put": {
        "tags": [
          "Policies"
        ],
        "summary": "Upgrade existing policy.",
        "description": "Upgrades policy to another policy type.",
        "operationId": "upgrade",
        "parameters": [
          {
            "name": "policyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradePolicyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradePolicyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradePolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "Policy upgraded successfully."
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Policy not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/policies/{policyId}/renew": {
      "post": {
        "tags": [
          "Policies"
        ],
        "operationId": "renew",
        "parameters": [
          {
            "name": "policyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/policies/{policyId}": {
      "get": {
        "tags": [
          "Policies"
        ],
        "operationId": "getDetails",
        "parameters": [
          {
            "name": "policyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePolicyTypeCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "annualCoverageLimit": {
            "type": "number",
            "format": "double"
          },
          "deductibleAmount": {
            "type": "number",
            "format": "double"
          },
          "coverageCategories": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EnrollPolicyCommand": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "policyTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
...
      "UpgradePolicyRequest": {
        "type": "object",
        "properties": {
          "newPolicyTypeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}